+++ extensions/lib/hello-someone/hello-someone.d.ts
@@ -1,4 +1,4 @@
-import type { IDataProvider } from 'shared/models/data-provider.model';
+import type IDataProvider from 'shared/models/data-provider.interface';
 
 export interface GreetingsDataProvider extends IDataProvider<string, string, string> {
   testRandomMethod(things: string): Promise<string>;
