+++ extensions/lib/quick-verse/quick-verse.d.ts
@@ -1,4 +1,4 @@
-import type { IDataProvider } from 'shared/models/data-provider.model';
+import type IDataProvider from 'shared/models/data-provider.interface';
 
 // TODO: Move these types to quick-verse.ts and generate this file from quick-verse.ts in the future?
