+++ src/renderer/hooks/papi-hooks/use-data-provider.hook.ts
@@ -1,5 +1,5 @@
 import dataProviderService from '@shared/services/data-provider.service';
-import { IDataProvider } from '@shared/models/data-provider.model';
+import IDataProvider from '@shared/models/data-provider.interface';
 import { useCallback, useMemo, useState } from 'react';
 import useEvent from '@renderer/hooks/papi-hooks/use-event.hook';
 import usePromise from '@renderer/hooks/papi-hooks/use-promise.hook';
