+++ src/renderer/hooks/papi-hooks/use-data.hook.ts
@@ -1,5 +1,5 @@
-import { DataProviderSubscriberOptions } from '@shared/models/data-provider.interface';
-import { IDataProvider } from '@shared/models/data-provider.model';
+import { DataProviderSubscriberOptions } from '@shared/models/data-provider.model';
+import IDataProvider from '@shared/models/data-provider.interface';
 import useEventAsync from '@renderer/hooks/papi-hooks/use-event-async.hook';
 import { useMemo, useState } from 'react';
 import { PapiEventAsync, PapiEventHandler } from '@shared/models/papi-event.model';
